@@ -1,6 +1,6 @@
 {
   "name": "kfk",
-  "version": "0.0.7",
+  "version": "0.1.0",
   "description": "The high-level node kafka client based on node-rdkafka  .",
   "main": "dist/src/index.js",
   "types": "dist/src/index.d.ts",
@@ -21,8 +21,8 @@
     "clean": "rm -rf dist",
     "purge": "rm -rf package-lock.json node_modules && npm run clean",
     "build": "npm run clean && npm run compile",
-    "lint": "tslint --project .",
-    "test": "npm run lint && npm run compile && NODE_ENV=ci ava dist/tests",
+    "lint": "tslint src/**/*.ts --project .",
+    "test": "npm run clean && npm run compile && NODE_ENV=ci ava dist/tests -v",
     "prettier": "prettier --config ./.prettierrc --write '{src,tests}/**/*.ts'",
     "pub": "npm run build && npm publish"
   },
@@ -31,14 +31,14 @@
   ],
   "author": "joway",
   "dependencies": {
-    "@types/bluebird": "^3.5.20",
-    "@types/lodash": "^4.14.106",
-    "bluebird": "^3.5.1",
+    "@types/bluebird": "3.5.20",
+    "@types/lodash": "4.14.106",
+    "bluebird": "3.5.1",
     "lodash": "4.17.5",
-    "node-rdkafka": "^2.3.1"
+    "node-rdkafka": "2.3.4"
   },
   "devDependencies": {
-    "@types/node": "8.0.46",
+    "@types/node": "^10.5.6",
     "@types/sinon": "2.3.6",
     "@types/supertest": "2.0.3",
     "ava": "0.24.0",
@@ -48,7 +48,8 @@
     "sinon": "4.0.1",
     "supertest": "3.0.0",
     "tslint": "5.9.1",
+    "tslint-config-airbnb": "^5.9.2",
     "tslint-jike-node": "0.0.14",
-    "typescript": "2.7.2"
+    "typescript": "2.9.2"
   }
 }
