@@ -1,23 +1,18 @@
-import * as _ from 'lodash'
 import * as Kafka from 'node-rdkafka'
-import * as bluebird from 'bluebird'
 
 import { KafkaProducer } from './producer'
 import { KafkaALOConsumer, KafkaAMOConsumer } from './consumer'
 import { KfkErrorCode } from './errors'
 import { TopicPartition, KafkaMetadata, KafkaMessage, KafkaMessageError } from './types'
 
-
 export {
+  Kafka,
   TopicPartition,
   KafkaMetadata,
   KafkaMessage,
   KafkaMessageError,
-
   KafkaProducer,
-
   KafkaALOConsumer,
   KafkaAMOConsumer,
-
   KfkErrorCode,
 }
